requirements_fixed.txt
asgiref==3.8.1
asttokens==2.4.1
blinker==1.9.0
boto3==1.38.5
botocore==1.38.5
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.1.6
django-allauth==65.7.0
django-appconf==1.1.0
django-auditlog==3.0.0
django-autoslug==1.9.9
django-betterforms==2.0.0
django-ckeditor==6.7.2
django-compressor==4.5.1
django-cors-headers==4.7.0
django-countries==7.6.1
django-crispy-forms==2.4
django-filter==25.1
django-haystack==3.3.0
django-heroku==0.3.1
django-js-asset==3.1.2
django-mapbox-location-field==2.1.0
django-otp==1.6.0
django-paypal==2.1
django-storages==1.14.6
django-tinymce==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djose==2.3.1
executing==2.1.0
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.57.0
greenlet==3.2.1
idna==3.10
instamojo-wrapper==1.2.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.5
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mysql==0.0.3
mysqlclient==2.2.7
nest-asyncio==1.6.0
numpy==2.2.5
oauthlib==3.2.2
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
pywin32==306
pyzmq==26.2.0
razorpay==1.4.2
rcssmin==1.1.2
requests==2.32.3
requests-oauthlib==2.0.0
rjsmin==1.2.2
s3transfer==0.12.0
setuptools==80.0.0
six==1.16.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.9.0
